Add draft case study for OCP TiogaPass
Following discussion in OCP OSF, Wiwynn makes a special OCP
TiogaPass SKU that enables software development community to
work on OSF solutions. Such SKU can be ordered with unit of
one server, and it can be used with or without a rack.

This commit adds a draft case study. To start order process,
Wiwynn needs a URL, as a landing page for software developers
to know how to use the SKU.

This draft case study only contains a landing page, with
skeleton of information to be provided.

We will refine this landing page and detail pages linked by it
over time.  They need to be good enough for consumption when Wiwynn
sells first order to the community, which is expected in July, 2019.

This landing page will be reviewed by Linuxboot community and Wiwynn.

Signed-off-by: Jonathan Zhang <jonzhang@fb.com>