src: Use user-space consistently
user-space is used by Red Hat and makes more sense than "user space" as
on Wikipedia. This also passes the vale linting if we ever enable it.

Remove the definition of user-space since it is better defined on
Wikipedia and there is nothing unique about the definition in our use
case.

Signed-off-by: Philip Molloy <philip@philipmolloy.com>